Add AABB for meshes and submeshes
Each mesh and submesh will contain their Axis Aligned Bounding Box,
which will allow certain optimizations.

New descriptors have been created to contain extra data about all the
instances and all the submeshes used in the current scene (including
their AABBs).

+++ Engine/src/AssetManager/AssetLoad/MaterialLoader.cpp
@@ -8,7 +8,7 @@
 namespace mtd
 {
     constexpr uint64_t MATERIAL_MAGIC = "MTD_MTRL"_u64;
-    constexpr uint32_t MATERIAL_FILE_VERSION = 1U;
+    constexpr uint64_t MATERIAL_FILE_VERSION = 1UL;
     constexpr size_t MATERIAL_ALIGNMENT = 16;
 }
